# bedelia-rest/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bedelia UTEC",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/edicionesCurso/inscripciones/{ciEstudiante}": {
            "post": {
                "tags": [
                    "Ediciones Curso"
                ],
                "description": "Inscribe a un estudiante a ediciones de curso",
                "operationId": "App\\Http\\Controllers\\EdicionesCursoController::asignarEstudiante",
                "parameters": [
                    {
                        "name": "ciEstudiante",
                        "in": "path",
                        "description": "CI del estudiante a inscribir",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "ID de los EdicionCurso a los que se inscribirá el estudiante",
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/edicionesCurso/{id}/docente/{ciDocente}": {
            "put": {
                "tags": [
                    "Ediciones Curso"
                ],
                "description": "Asigna un docente a una edicion de curso",
                "operationId": "App\\Http\\Controllers\\EdicionesCursoController::asignarDocente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del edicion curso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ciDocente",
                        "in": "path",
                        "description": "CI del docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
        "/edicionesCurso/{id}/estudiantes": {
            "get": {
                "tags": [
                    "Ediciones Curso"
                ],
                "description": "Devuelve los estudiantes inscritos al EdicionCurso",
                "operationId": "App\\Http\\Controllers\\EdicionesCursoController::estudiantes",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del edicion curso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClaseDictadaDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/edicionesCurso/{ciEstudiante}/{idCarrera}": {
            "get": {
                "tags": [
                    "Ediciones Curso"
                ],
                "description": "Los EdicionCurso a los que el estudiante puede inscribirse en el PeriodoLectivo actual",
                "operationId": "App\\Http\\Controllers\\EdicionesCursoController::listarParaInscripcion",
                "parameters": [
                    {
                        "name": "ciEstudiante",
                        "in": "path",
                        "description": "CI del estudiante",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idCarrera",
                        "in": "path",
                        "description": "ID de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EdicionCursoDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/edicionesCurso/{id}/notas": {
            "get": {
                "tags": [
                    "Ediciones Curso"
                ],
                "description": "Obtiene las notas obtenidas por los estudiantes",
                "operationId": "App\\Http\\Controllers\\EdicionesCursoController::ObtenerNotas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del edicion curso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActaDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Ediciones Curso"
                ],
                "description": "Marca como confirmada el acta del EdicionCurso",
                "operationId": "App\\Http\\Controllers\\EdicionesCursoController::ConfirmarActa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del edicion curso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActaDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ediciones Curso"
                ],
                "description": "Registra las notas obtenidas por los estudiantes",
                "operationId": "App\\Http\\Controllers\\EdicionesCursoController::IngresarNotas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del edicion curso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActaDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActaDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/examenes/inscripciones/{ciEstudiante}": {
            "post": {
                "tags": [
                    "Exámenes"
                ],
                "description": "Inscribe a un estudiante a exámenes",
                "operationId": "App\\Http\\Controllers\\ExamenController::asignarEstudiante",
                "parameters": [
                    {
                        "name": "ciEstudiante",
                        "in": "path",
                        "description": "CI del estudiante a inscribir",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "ID de los Examenes a los que se inscribirá el estudiante",
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/examenes/{id}/docente/{ciDocente}": {
            "put": {
                "tags": [
                    "Exámenes"
                ],
                "description": "Asigna un docente a una exámen",
                "operationId": "App\\Http\\Controllers\\ExamenController::asignarDocente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del exámen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ciDocente",
                        "in": "path",
                        "description": "CI del docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/examenes/docente/{ciDocente}": {
            "get": {
                "tags": [
                    "Exámenes"
                ],
                "description": "devuelve los Examen que el docente toma en el PeriodoExamen actual",
                "operationId": "App\\Http\\Controllers\\ExamenController::ExamenessDocente",
                "parameters": [
                    {
                        "name": "ciDocente",
                        "in": "path",
                        "description": "CI del docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExamenDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/examenes/{ciEstudiante}/{idCarrera}": {
            "get": {
                "tags": [
                    "Exámenes"
                ],
                "description": "Devuelve los Examenes a los que el estudiante puede inscribirse en el PeriodoExamen actual",
                "operationId": "App\\Http\\Controllers\\ExamenController::listarParaInscripcion",
                "parameters": [
                    {
                        "name": "ciEstudiante",
                        "in": "path",
                        "description": "CI del estudiante",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idCarrera",
                        "in": "path",
                        "description": "ID de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExamenDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EdicionCursoDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "acta_confirmada": {
                        "type": "boolean"
                    },
                    "habilitado": {
                        "description": "1 = el estudiante está habilitado para inscribirse, 0 = curso ya aprobado, -1 = no se cumple con las previas",
                        "type": "integer",
                        "readOnly": true
                    },
                    "sede": {
                        "$ref": "#/components/schemas/SedeDTO"
                    },
                    "curso": {
                        "$ref": "#/components/schemas/CursoDTO"
                    },
                    "docente": {
                        "$ref": "#/components/schemas/UsuarioDTO"
                    }
                },
                "type": "object"
            },
            "ExamenDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "fecha": {
                        "type": "string"
                    },
                    "acta_confirmada": {
                        "type": "boolean"
                    },
                    "habilitado": {
                        "description": "1 = el estudiante está habilitado para inscribirse, 0 = no es necesario dar el examen, -1 = no ha ganado el derecho a dar examen",
                        "type": "integer",
                        "readOnly": true
                    },
                    "sede": {
                        "$ref": "#/components/schemas/SedeDTO"
                    },
                    "curso": {
                        "$ref": "#/components/schemas/CursoDTO"
                    },
                    "docente": {
                        "$ref": "#/components/schemas/UsuarioDTO"
                    }
                },
                "type": "object"
            },
            "ClaseDictadaDTO": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "fecha": {
                        "type": "string",
                        "readOnly": true
                    },
                    "curso_id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "edicion_curso_id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "lista": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "ciEstudiante": {
                                    "type": "string"
                                },
                                "nombre": {
                                    "type": "string",
                                    "readOnly": true
                                },
                                "apellido": {
                                    "type": "string",
                                    "readOnly": true
                                },
                                "asistencia": {
                                    "type": "number",
                                    "writeOnly": true
                                },
                                "total_asistencias": {
                                    "type": "number",
                                    "readOnly": true
                                },
                                "total_llegadas_tarde": {
                                    "type": "number",
                                    "readOnly": true
                                },
                                "total_inasistencias": {
                                    "type": "number",
                                    "readOnly": true
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Token de autenticacion conformato como '<token>'",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}